+++ NEWS.md
@@ -31,3 +31,11 @@
 -   Fixed - bugs
 -   Updated - vignettes & online tutorials
 -   Updated - Documents & references
+
+# LAWBL 1.5.0
+
+-   Updated - pcfa,pefa,pcirm,sim_lvm,summary
+-   Updated - sim_lvm,summary.lawbl,plot_lawbl
+-   Fixed - bugs
+-   Updated - vignettes & online tutorials
+-   Updated - Documents & references